requirements.txt
paho-mqtt==2.1.0
protobuf==5.27.1
